requirements.txt
altair==4.2.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
color-matcher==0.5.0
ddt==1.6.0
decorator==5.1.1
docutils==0.20
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
instagrapi==1.17.8
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
pyarrow==12.0.0
pycryptodomex==3.17
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
streamlit==1.22.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
